Create a data module to hide all data related code
