+++ .idea/modules.xml
@@ -3,7 +3,10 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/CinemaBuff.iml" filepath="$PROJECT_DIR$/CinemaBuff.iml" />
+      <module fileurl="file://$PROJECT_DIR$/buildSrc/CinemaBuff-buildSrc.iml" filepath="$PROJECT_DIR$/buildSrc/CinemaBuff-buildSrc.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/buildSrc/buildSrc.iml" filepath="$PROJECT_DIR$/buildSrc/buildSrc.iml" />
+      <module fileurl="file://$PROJECT_DIR$/data/data.iml" filepath="$PROJECT_DIR$/data/data.iml" />
     </modules>
   </component>
 </project>
