+++ app/src/main/java/com/levimoreira/cinemabuff/movie/data/MovieDataSource.kt
@@ -2,53 +2,52 @@ package com.levimoreira.cinemabuff.movie.data
 
 import androidx.lifecycle.MutableLiveData
 import androidx.paging.PageKeyedDataSource
-import com.levimoreira.cinemabuff.infrastructure.data.NetworkState
-import com.levimoreira.cinemabuff.infrastructure.data.models.Movie
+import com.cinemabuff.data.network.NetworkState
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.schedulers.Schedulers
 import timber.log.Timber
 
-class MovieDataSource(private val repository: MovieRepository) : PageKeyedDataSource<Int, Movie>() {
+class MovieDataSource(private val repository: MovieRepository) : PageKeyedDataSource<Int, com.cinemabuff.data.entities.Movie>() {
 
     private val disposableBag = CompositeDisposable()
     var loadState: MutableLiveData<NetworkState> = MutableLiveData()
 
     override fun loadInitial(
-        params: LoadInitialParams<Int>,
-        callback: LoadInitialCallback<Int, Movie>
+            params: LoadInitialParams<Int>,
+            callback: LoadInitialCallback<Int, com.cinemabuff.data.entities.Movie>
     ) {
         loadState.postValue(NetworkState.LOADING)
         val disposable = repository.getMovies(page = 1)
-            .subscribeOn(Schedulers.io())
-            .observeOn(AndroidSchedulers.mainThread())
-            .subscribe({
-                callback.onResult(it.results, null, 2)
-                loadState.postValue(NetworkState.LOADED)
-            }, {
-                Timber.e(it)
-                loadState.postValue(NetworkState.error(it.message))
-            })
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe({
+                    callback.onResult(it.results, null, 2)
+                    loadState.postValue(NetworkState.LOADED)
+                }, {
+                    Timber.e(it)
+                    loadState.postValue(NetworkState.error(it.message))
+                })
 
         disposableBag.add(disposable)
     }
 
-    override fun loadAfter(params: LoadParams<Int>, callback: LoadCallback<Int, Movie>) {
+    override fun loadAfter(params: LoadParams<Int>, callback: LoadCallback<Int, com.cinemabuff.data.entities.Movie>) {
         loadState.postValue(NetworkState.LOADING)
         val disposable = repository.getMovies(page = params.key)
-            .subscribeOn(Schedulers.io())
-            .observeOn(AndroidSchedulers.mainThread())
-            .subscribe({
-                callback.onResult(it.results, params.key + 1)
-                loadState.postValue(NetworkState.LOADED)
-            }, {
-                Timber.e(it)
-                loadState.postValue(NetworkState.error(it.message))
-            })
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe({
+                    callback.onResult(it.results, params.key + 1)
+                    loadState.postValue(NetworkState.LOADED)
+                }, {
+                    Timber.e(it)
+                    loadState.postValue(NetworkState.error(it.message))
+                })
         disposableBag.add(disposable)
     }
 
-    override fun loadBefore(params: LoadParams<Int>, callback: LoadCallback<Int, Movie>) {
+    override fun loadBefore(params: LoadParams<Int>, callback: LoadCallback<Int, com.cinemabuff.data.entities.Movie>) {
         loadState.postValue(NetworkState.LOADED)
     }
