+++ app/src/main/java/com/levimoreira/cinemabuff/movie/data/MovieDataSourceFactory.kt
@@ -1,10 +1,10 @@
 package com.levimoreira.cinemabuff.movie.data
 
 import androidx.paging.DataSource
-import com.levimoreira.cinemabuff.infrastructure.data.models.Movie
+import com.cinemabuff.data.entities.Movie
 
 class MovieDataSourceFactory(private val movieRepository: MovieRepository) :
-    DataSource.Factory<Int, Movie>() {
+        DataSource.Factory<Int, Movie>() {
     override fun create(): DataSource<Int, Movie> {
         return MovieDataSource(movieRepository)
     }
