+++ app/src/main/java/com/levimoreira/cinemabuff/movie/data/MovieRepository.kt
@@ -1,6 +1,6 @@
 package com.levimoreira.cinemabuff.movie.data
 
-import com.levimoreira.cinemabuff.infrastructure.data.retrofit.TopMovieListResponse
+import com.cinemabuff.data.models.TopMovieListResponse
 import io.reactivex.Single
 
 interface MovieRepository {
