+++ app/src/main/java/com/levimoreira/cinemabuff/movie/data/MovieRepositoryImpl.kt
@@ -1,11 +1,11 @@
 package com.levimoreira.cinemabuff.movie.data
 
-import com.levimoreira.cinemabuff.infrastructure.api.TmdbApi
-import com.levimoreira.cinemabuff.infrastructure.data.retrofit.TopMovieListResponse
+import com.cinemabuff.data.models.TopMovieListResponse
+import com.cinemabuff.data.network.TmdbApi
 import io.reactivex.Single
 
 class MovieRepositoryImpl(val tmdbService: TmdbApi) :
-    MovieRepository {
+        MovieRepository {
 
     override fun getMovies(page: Int): Single<TopMovieListResponse> {
         return tmdbService.getTopMovies(page = page)
