+++ app/src/main/java/com/levimoreira/cinemabuff/movie/viewmodels/MovieViewModel.kt
@@ -4,24 +4,24 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.paging.LivePagedListBuilder
 import androidx.paging.PagedList
-import com.levimoreira.cinemabuff.infrastructure.api.TmdbApi.Companion.TMDB_PAGE_SIZE
-import com.levimoreira.cinemabuff.infrastructure.data.models.Movie
+import com.cinemabuff.data.entities.Movie
+import com.cinemabuff.data.network.TmdbApi.Companion.TMDB_PAGE_SIZE
 import com.levimoreira.cinemabuff.movie.data.MovieDataSource
 import com.levimoreira.cinemabuff.movie.data.MovieDataSourceFactory
 
 class MovieViewModel(dataSourceFactory: MovieDataSourceFactory) : ViewModel() {
 
     private val config = PagedList.Config.Builder()
-        .setPageSize(TMDB_PAGE_SIZE)
-        .build()
+            .setPageSize(TMDB_PAGE_SIZE)
+            .build()
 
     val movieDataSource: MovieDataSource = dataSourceFactory.create() as MovieDataSource
 
     private val pagedList = LivePagedListBuilder(dataSourceFactory, config)
-        .setInitialLoadKey(1)
-        .build()
+            .setInitialLoadKey(1)
+            .build()
 
-    fun getMovies(): LiveData<PagedList<Movie>>{
+    fun getMovies(): LiveData<PagedList<Movie>> {
         return pagedList
     }
