+++ app/src/main/java/com/levimoreira/cinemabuff/movie/views/MovieListAdapter.kt
@@ -5,21 +5,21 @@ import android.view.ViewGroup
 import androidx.paging.PagedListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
+import com.cinemabuff.data.entities.Movie
 import com.levimoreira.cinemabuff.R
-import com.levimoreira.cinemabuff.infrastructure.data.models.Movie
 import com.levimoreira.cinemabuff.movie.data.MovieDiffUtil
 import com.levimoreira.cinemabuff.utils.inflate
 import kotlinx.android.synthetic.main.movie_list_item.view.*
 
 class MovieListAdapter :
-    PagedListAdapter<Movie, MovieListAdapter.MovieViewHolder>(
-        MovieDiffUtil()
-    ) {
+        PagedListAdapter<Movie, MovieListAdapter.MovieViewHolder>(
+                MovieDiffUtil()
+        ) {
 
     lateinit var onMovieClickListener: (Movie) -> Unit
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MovieViewHolder =
-        MovieViewHolder(parent.inflate(R.layout.movie_list_item))
+            MovieViewHolder(parent.inflate(R.layout.movie_list_item))
 
     override fun onBindViewHolder(holder: MovieViewHolder, position: Int) {
         getItem(position)?.let {
@@ -29,8 +29,8 @@ class MovieListAdapter :
 
     inner class MovieViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
         fun bind(
-            item: Movie,
-            onMovieClickListener: (Movie) -> Unit
+                item: Movie,
+                onMovieClickListener: (Movie) -> Unit
         ) = with(itemView) {
             Glide.with(itemView).load(item.getCompletePosterPath()).into(moviePoster)
             movieTitle.text = item.title
