+++ build.gradle
@@ -2,6 +2,35 @@
 
 
 buildscript {
+    ext.kotlin_version = '1.3.30'
+
+    ext.versions = [
+            room    : '2.1.0-alpha04',
+            retrofit: '2.4.0',
+            gson    : '2.8.5',
+            koin    : '1.0.2'
+    ]
+
+    ext.deps = [
+            room    : [
+                    'runtime'   : "androidx.room:room-runtime:${versions.room}",
+                    'compiler'  : "androidx.room:room-compiler:${versions.room}",
+                    'rxjava2'   : "androidx.room:room-rxjava2:${versions.room}",
+                    'coroutines': "androidx.room:room-coroutines:${versions.room}"
+            ],
+
+            retrofit: [
+                    'core'          : "com.squareup.retrofit2:retrofit:${versions.retrofit}",
+                    'gsonconverter' : "com.squareup.retrofit2:converter-gson:${versions.retrofit}",
+                    'rxjava2adapter': "com.squareup.retrofit2:adapter-rxjava2:${versions.retrofit}"
+            ],
+            gson    : "com.google.code.gson:gson:${versions.gson}",
+            koin    : [
+                    'core'     : "org.koin:koin-android:${versions.koin}",
+                    'viewmodel': "org.koin:koin-android-viewmodel:${versions.koin}"
+            ]
+    ]
+
 
     repositories {
         google()
@@ -9,13 +38,14 @@ buildscript {
     }
     dependencies {
 
-        classpath 'com.android.tools.build:gradle:3.3.1'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.3.21"
+        classpath 'com.android.tools.build:gradle:3.5.0-beta01'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.3.31"
         classpath 'com.novoda:gradle-build-properties-plugin:0.4.1'
 
 
         // NOTE: Do not place your application versions here; they belong
         // in the individual module build.gradle files
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
