+++ data/src/main/java/com/cinemabuff/data/database/MovieDao.kt
@@ -1,9 +1,9 @@
-package com.levimoreira.cinemabuff.infrastructure.database
+package com.cinemabuff.data.database
 
 import androidx.lifecycle.LiveData
 import androidx.room.Dao
 import androidx.room.Query
-import com.levimoreira.cinemabuff.infrastructure.data.models.Movie
+import com.cinemabuff.data.entities.Movie
 
 @Dao
 interface MovieDao : BaseDao<Movie> {
