+++ data/src/main/java/com/cinemabuff/data/network/ApiKeyInterceptor.kt
@@ -1,6 +1,6 @@
-package com.levimoreira.cinemabuff.infrastructure.api
+package com.cinemabuff.data.network
 
-import com.levimoreira.cinemabuff.BuildConfig
+import com.cinemabuff.data.BuildConfig
 import okhttp3.Interceptor
 import okhttp3.Response
 
@@ -10,12 +10,12 @@ class ApiKeyInterceptor : Interceptor {
         val originalHttpUrl = original.url()
 
         val url = originalHttpUrl.newBuilder()
-            .addQueryParameter(API_KEY_PARAM, BuildConfig.TMDB_API_KEY)
-            .build()
+                .addQueryParameter(API_KEY_PARAM, BuildConfig.TMDB_API_KEY)
+                .build()
 
         // Request customization: add request headers
         val requestBuilder = original.newBuilder()
-            .url(url)
+                .url(url)
 
         val request = requestBuilder.build()
         return chain.proceed(request)
